requirements.txt
Django==1.11.4
django-filters==0.2.1
djangorestframework==3.7.1
Markdown==2.6.9
pkg-resources==0.0.0
pytz==2017.2
